Fix i18n loader circular dependencies
