+++ packages/sdk-react/src/utils/test-utils.tsx
@@ -1,11 +1,9 @@
 import { createAPIClient } from '@/api/client';
 import { getDefaultComponentSettings } from '@/core/componentSettings';
+import { getLocaleWithDefaults } from '@/core/context/I18nLoader';
 import { MoniteAPIProvider } from '@/core/context/MoniteAPIProvider';
 import { MoniteContext, MoniteTheme } from '@/core/context/MoniteContext';
-import {
-  getLocaleWithDefaults,
-  MoniteI18nProvider,
-} from '@/core/context/MoniteI18nProvider';
+import { MoniteI18nProvider } from '@/core/context/MoniteI18nProvider';
 import {
   MoniteProviderProps,
   MoniteSettings,
